JWWEB-737 Moved subcomponents of Transaction Details to components directory
