+++ src/components/TransactionDetails/components/AssetItemPreview/assetItemPreview.m.scss
@@ -1,19 +1,12 @@
 @import 'styles/mixins';
 
 .core {
-  @include property-transition(box-shadow);
+  @include flex-row;
 
-  display: flex;
-  width: 100%;
-  height: 72px;
-  max-height: 72px;
+  position: relative;
+  padding: 14px 0 16px;
   justify-content: space-between;
-  align-items: stretch;
-  flex-direction: row;
-  flex-wrap: wrap;
-  border: none;
   background-color: $white;
-  will-change: height, opacity;
 
   //&:focus,
   &:hover {
@@ -26,34 +19,38 @@
 }
 
 .item {
-  @include font(16px, 24px);
-
   display: flex;
   flex-direction: column;
   position: relative;
   justify-content: center;
   align-items: center;
-  font-weight: 600;
-  color: $main-black;
   min-width: 64px;
 }
 
-.text {
-  margin-bottom: 2px;
+.data {
+  overflow: hidden;
 }
 
-.subtext {
-  @include font(14px, 22px);
+.label {
+  @include ellipsis;
 
-  font-weight: 400;
+  font-size: 12px;
+  line-height: 20px;
   color: $main-black-40;
+}
+
+.body {
+  @include ellipsis;
+
+  font-size: 14px;
+  font-weight: 600;
+  line-height: 22px;
+  color: $main-black;
 }
 
 .main-block {
   align-items: flex-start;
   flex-grow: 1;
-  white-space: nowrap;
-  text-overflow: ellipsis;
 }
 
 .arrow-icon,
