+++ src/components/TransactionDetails/components/FieldPreview/fieldPreview.m.scss
