+++ src/components/TransactionDetails/components/index.js
@@ -1,0 +1,4 @@
+// @flow strict
+
+export { AssetItemPreview } from './AssetItemPreview/AssetItemPreview'
+export { FieldPreview } from './FieldPreview/FieldPreview'
