+++ src/components/index.js
@@ -36,7 +36,6 @@ import DigitalAssetEditForm from './DigitalAssetEditForm'
 export { TitleHeader } from './TitleHeader/TitleHeader'
 export { PasswordInput } from './PasswordInput/PasswordInput'
 export { NewPasswordField } from './NewPasswordField/NewPasswordField'
-export { FieldPreview } from './FieldPreview/FieldPreview'
 
 export {
   QRCode,
